Separate NPE into its own module
This just removes the Naomi Plugin Exchange functions from
application.py and moves them into their own module (npe.py).
This makes application.py smaller and easier to work with and
should make it easier to troubleshoot NPE issues.
