+++ naomi/__main__.py
@@ -4,18 +4,33 @@ import logging
 import argparse
 from . import application
 from . import app_utils
-from . import profile
 from . import coloredformatting as cf
+from . import i18n
+from . import paths
+from . import profile
+
+
 logo = cf.naomidefaults.logo
 sto = cf.naomidefaults.sto
-
-
 USE_STANDARD_MIC = application.USE_STANDARD_MIC
 USE_TEXT_MIC = application.USE_TEXT_MIC
 USE_BATCH_MIC = application.USE_BATCH_MIC
 
 
 def main(args=None):
+    logger = logging.getLogger(__name__)
+    language = profile.get_profile_var(['language'])
+    if(not language):
+        language = 'en-US'
+        logger.warn(
+            ' '.join([
+                'language not specified in profile,',
+                'using default ({})'.format(language)
+            ])
+        )
+    translations = i18n.parse_translations(paths.data('locale'))
+    translator = i18n.GettextMixin(translations)
+    _ = translator.gettext
     parser = argparse.ArgumentParser(description='Naomi Voice Control Center')
     parser.add_argument(
         '--debug',
@@ -125,7 +140,7 @@ def main(args=None):
     mic_mode.add_argument(
         '--local',
         action='store_true',
-        help='Use text input instead of a real microphone'
+        help='Use text input/output instead of verbal interface'
     )
     mic_mode.add_argument(
         '--batch',
@@ -133,8 +148,7 @@ def main(args=None):
         metavar="FILE",
         type=argparse.FileType('r'),
         help=' '.join([
-            'Batch mode using a text file with text',
-            'commands audio filenames at each line.'
+            'Batch mode using a text file with references to audio files'
         ])
     )
     mic_mode.add_argument(
@@ -150,12 +164,12 @@ def main(args=None):
     print("    /::|  |       /::\  \       /::\  \       /::|  |        /\  \   ")
     print("   /:|:|  |      /:/\:\  \     /:/\:\  \     /:|:|  |        \:\  \  ")
     print("  /:/|:|  |__   /::\~\:\  \   /:/  \:\  \   /:/|:|__|__      /::\__\ ")
-    print(" /:/ |:| /\__\ /:/\:\ \:\__\ /:/__/ \:\__\ /:/ |::::\__\  __/:/\/__/ ")
-    print(" \/__|:|/:/  / \/__\:\/:/  / \:\  \ /:/  / \/__/~~/:/  / /\/:/  /    ")
-    print("     |:/:/  /       \::/  /   \:\  /:/  /        /:/  /  \::/__/     ")
-    print("     |::/  /        /:/  /     \:\/:/  /        /:/  /    \:\__\     ")
-    print("     /:/  /        /:/  /       \::/  /        /:/  /      \/__/     ")
-    print("     \/__/         \/__/         \/__/         \/__/                 ")
+    print(" /:/ |:| /\__\ /:/\:\ \:\__\ /:/__/ \:\__\ /:/ |::::\__\    /:/\/__/ ")
+    print(" \/__|:|/:/  / \/__\:\/:/  / \:\  \ /:/  / \/__/~~/:/  / __/:/  /    ")
+    print("     |:/:/  /       \::/  /   \:\  /:/  /        /:/  / /\/:/  /     ")
+    print("     |::/  /        /:/  /     \:\/:/  /        /:/  /  \::/__/      ")
+    print("     /:/  /        /:/  /       \::/  /        /:/  /    \:\__\      ")
+    print("     \/__/         \/__/         \/__/         \/__/      \/__/      ")
     print(sto)
 
     # Set up logging
@@ -207,29 +221,47 @@ def main(args=None):
         save_active_audio=p_args.save_active_audio,
         save_noise=p_args.save_noise
     )
-    if p_args.list_active_plugins:
-        app.list_active_plugins()
-        sys.exit(0)
-    elif p_args.list_audio_devices:
+    if p_args.list_audio_devices:
         app.list_audio_devices()
         sys.exit(0)
+    if p_args.list_active_plugins:
+        print(_("Active Plugins:"))
+        active_plugins = app.npe.list_active_plugins()
+        len_name = max(len(active_plugins[info].name) for info in active_plugins)
+        len_version = max(len(active_plugins[info].version) for info in active_plugins)
+        for name in sorted(active_plugins):
+            info = active_plugins[name]
+            print(
+                "{} {} - {}".format(
+                    info.name.ljust(len_name),
+                    ("(v%s)" % info.version).ljust(len_version),
+                    info.description
+                )
+            )
+        sys.exit(0)
     if p_args.list_available:
-        app.list_available_plugins(p_args.list_available)
+        print(_("Available Plugins:"))
+        print_plugins = app.npe.list_available_plugins(p_args.list_available)
+        if(len(print_plugins) == 0):
+            print(_("Sorry, no plugins matched"))
+        else:
+            for name in sorted(print_plugins):
+                print(print_plugins[name])
         sys.exit(0)
     if p_args.plugins_to_install:
-        app.install_plugins(p_args.plugins_to_install)
+        print(app.npe.install_plugins(p_args.plugins_to_install))
         sys.exit(0)
     if p_args.plugins_to_update:
-        app.update_plugins(p_args.plugins_to_update)
+        print(app.npe.update_plugins(p_args.plugins_to_update))
         sys.exit(0)
     if p_args.plugins_to_remove:
-        app.remove_plugins(p_args.plugins_to_remove)
+        print(app.npe.remove_plugins(p_args.plugins_to_remove))
         sys.exit(0)
     if p_args.plugins_to_enable:
-        app.enable_plugins(p_args.plugins_to_enable)
+        print(app.npe.enable_plugins(p_args.plugins_to_enable))
         sys.exit(0)
     if p_args.plugins_to_disable:
-        app.disable_plugins(p_args.plugins_to_disable)
+        print(app.npe.disable_plugins(p_args.plugins_to_disable))
         sys.exit(0)
     app.run()
